added comments and script descriptions
